mcp server dev, and cli interface and restructure codebase
