+++ src/ai/index.ts
@@ -1,0 +1,6 @@
+// AI/ML Module
+export * from './semantic-search.js';
+export * from './embedding-service.js';
+export * from './embedding-optimizer.js';
+export * from './vector-store.js';
+export * from './chroma-vector-store.js';
