+++ src/analysis/index.ts
@@ -1,0 +1,3 @@
+// Code Analysis Module
+export * from './code-parser.js';
+export * from './cross-reference-analyzer.js';
