+++ src/mcp/index.ts
@@ -1,0 +1,5 @@
+// MCP Server Module
+export * from './mcp-server.js';
+export * from './memory-manager.js';
+export * from './memory-config.js';
+export * from './version.js';
