+++ src/mcp/memory-config.ts
