+++ src/mcp/memory-manager.ts
