+++ src/mcp/version.ts
