+++ test/test-file-watcher.ts
@@ -1,6 +1,6 @@
 import { test, describe } from 'node:test';
 import assert from 'node:assert';
-import { FileWatcher } from '../src/file-watcher.js';
+import { FileWatcher } from '../src/storage/file-watcher.js';
 import fs from 'fs/promises';
 import path from 'path';
